Trabajo Victor Algarín 29/10/2013

+++ DesignerRHN/DesignerRHN-ejb/src/java/Entidades/Periodicidades.java
@@ -6,6 +6,7 @@ package Entidades;
 
 import java.io.Serializable;
 import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.Collection;
 import javax.persistence.Basic;
 import javax.persistence.Column;
@@ -42,7 +43,7 @@ public class Periodicidades implements Serializable {
     @Basic(optional = false)
     @NotNull
     @Column(name = "SECUENCIA")
-    private BigDecimal secuencia;
+    private BigInteger secuencia;
     @Basic(optional = false)
     @NotNull
     @Column(name = "CODIGO")
@@ -65,20 +66,20 @@ public class Periodicidades implements Serializable {
     public Periodicidades() {
     }
 
-    public Periodicidades(BigDecimal secuencia) {
+    public Periodicidades(BigInteger secuencia) {
         this.secuencia = secuencia;
     }
 
-    public Periodicidades(BigDecimal secuencia, short codigo) {
+    public Periodicidades(BigInteger secuencia, short codigo) {
         this.secuencia = secuencia;
         this.codigo = codigo;
     }
 
-    public BigDecimal getSecuencia() {
+    public BigInteger getSecuencia() {
         return secuencia;
     }
 
-    public void setSecuencia(BigDecimal secuencia) {
+    public void setSecuencia(BigInteger secuencia) {
         this.secuencia = secuencia;
     }
